+++ NadekoBot/Modules/Pokemon/PokeStats.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace NadekoBot.Modules.Pokemon
+{
+    class PokeStats
+    {
+    }
+}
