+++ NadekoBot/NadekoBot.csproj
@@ -188,7 +188,26 @@
     <Compile Include="Modules\NSFW.cs" />
     <Compile Include="Modules\Permissions.cs" />
     <Compile Include="Commands\RatelimitCommand.cs" />
-    <Compile Include="Modules\PokemonModule.cs" />
+    <Compile Include="Modules\Pokemon\PokemonModule.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\PokeType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\PsychicType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\BugType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\RockType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\GhostType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\Dragontype.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\DarkType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\SteelType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\WaterType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\ElectricType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\GrassType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\IceType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\FightingType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\PoisonType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\GroundType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\FlyingType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\FireType.cs" />
+    <Compile Include="Modules\Pokemon\PokemonTypes\NormalType.cs" />
+    <Compile Include="Modules\Pokemon\PokeStats.cs" />
     <Compile Include="Modules\Searches.cs" />
     <Compile Include="Modules\Translator\Helpers\GoogleTranslator.cs" />
     <Compile Include="Modules\Translator\TranslateCommand.cs" />
